feat(thermal): Int16 temps, real-world °C range -273..5000, calibrated elements
Storage:
- TemperatureField.temps: Int8Array → Int16Array (buffer: Int16 → Int32)
- ThermalLookups / VisualLookups temperature arrays: Int8 → Int16 / Uint16
- TEMP_MIN = -273, TEMP_MAX = 5000 as shared constants
- Serializer bumps to VERSION 2 (Int16 temps); v1 payloads still load and widen

Calibration — temperatures now in °C tied to real-world milestones:
- water freezes 0°, boils 100°
- oil ignites 275°, wood 300°, coal 400°, methane 537°
- gunpowder detonates 170°, nitro 220°, thermite 450°
- lava emits at 1100°, fire at 900°, torch at 1000°, spark 300°
- stone melts to lava 1200°; sand to glass 1700°; obsidian to lava 1000°
- mud dries past 100°; wax melts 60°; waxliq freezes back 50°
- cryo (liquid N2) at -196°, boils at -180°
- mercury now boils at 357°; uranium critical-mass 500°

Renderer:
- Heatmap LUT rebuilt against real-°C stops (navy at -273, neutral at 20,
  orange at 400-800, red 1200, yellow-white 2500, blue-white 4500+)
- LUT widened to 512 entries; indexed by (t - TEMP_MIN) * SCALE
- Tint-mode thresholds scaled for the new range (15° vs 60° gates)
- tempTint warm ramp saturates at 1500° (white-hot), cold at -200°
- Glow / cold visual configs updated with realistic onset temps

UI:
- Ambient slider range -200..5000 with 10° step (was -100..127 step 1)
- Default ambient 20° (room temperature), reset snaps back to 20°
- Readout colour adds white-hot tier for ≥1000°
- Thermometer tooltip documents the range

Tests:
- Unit clamp tests updated to TEMP_MIN/TEMP_MAX
- E2e spawn-temp: expects lava 1100, fire 800, cryo -196, etc.
- Ambient-coupling / heatmap-render: default 20°, test drive 800°

+++ src/core/Grid.ts
@@ -26,7 +26,7 @@ export class Grid implements GridApi {
    * falling through warm air doesn't instantly reheat at its destination
    * while leaving a phantom cold spot behind. Set via `linkField()`.
    */
-  private field: { temps: Int8Array } | null = null;
+  private field: { temps: Int16Array } | null = null;
 
   constructor(
     public readonly width: number,
@@ -83,7 +83,7 @@ export class Grid implements GridApi {
   }
 
   /** Attach a temperature field so swaps move temperatures along with cells. */
-  linkField(field: { temps: Int8Array }): void {
+  linkField(field: { temps: Int16Array }): void {
     this.field = field;
   }
